data_processing/pyproject.toml
[project]
name = "data-processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "networkx>=3.6.1",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "slovakrailways>=0.1.0",
]
